+++ packages/terra-clinical-result/src/ClinicalResultBloodPressure.jsx
@@ -213,7 +213,7 @@ const ClinicalResultBloodPressure = (props) => {
           >
             <Observation
               key={`Observation-Diastolic-${diastolic.eventId}`}
-              className={!compareStatusIsInError.diastolic && diastolic.interpretation ? cx('diastolic-intrepretation') : null}
+              className={cx({ 'diastolic-interpretation': !compareStatusIsInError.diastolic && diastolic.interpretation })}
               eventId={diastolic.eventId}
               result={diastolic.result}
               interpretation={!compareStatusIsInError.diastolic ? diastolic.interpretation : null}